fix comments from @liuzicheng1987
