+++ tests/json/test_inheritance2.hpp
@@ -1,0 +1,3 @@
+namespace test_inheritance2 {
+void test();
+}
