+++ tests/json/test_view.cpp
@@ -19,25 +19,6 @@ struct Person {
   int age;
 };
 
-struct EmptyBase1 {};
-struct EmptyBase2 {};
-struct Derived1 : public EmptyBase1 {
-  int x;
-  int y;
-};
-struct Derived2 : public EmptyBase1, public EmptyBase2 {
-  int x;
-  int y;
-  int z;
-};
-
-struct BaseX {
-  int x;
-  int y;
-};
-struct EmptyDerived :public BaseX {};
-
-
 void test() {
   std::cout << std::source_location::current().function_name() << std::endl;
 
@@ -51,16 +32,5 @@ void test() {
 
   write_and_read(lisa,
                  R"({"first_name":"Maggie","last_name":"Simpson","age":0})");
-
-  Derived1 derived1;
-  const auto derived1_view = rfl::to_view(derived1);
-  static_assert(derived1_view.size() == 2);
-  Derived2 derived2;
-  const auto derived2_view = rfl::to_view(derived2);
-  static_assert(derived2_view.size() == 3);
-  
-  EmptyDerived empty_derived;
-  auto empty_derived_view = rfl::to_view(empty_derived);
-  static_assert(empty_derived_view.size() == 2);
 }
 }  // namespace test_view
